azd_utils.py
